# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Museum App",
    "description": "Museum App API",
    "contact": {
      "name": "Mehmet Fatih Erdem",
      "url": "https://github.com/mehmetfatiherdem",
      "email": "mehmeterdem1024@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Localhost"
    },
    {
      "url": "https://ronchon-gerard-51245.herokuapp.com/",
      "description": "Heroku"
    }
  ],
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "tags": ["User"],
        "summary": "Get user",
        "description": "Get user",
        "operationId": "getUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "example": "62d06853e1dede40af51718e"
          }
        ],
        "responses": {
          "200": {
            "description": "Get user"
          }
        }
      }
    },
    "/api/comments/add": {
      "post": {
        "tags": ["Comment"],
        "summary": "Add comment",
        "description": "Add comment",
        "operationId": "addComment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "museumId": {
                    "type": "string",
                    "required": true
                  },
                  "text": {
                    "type": "string",
                    "required": true,
                    "example": "This is a comment"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "authCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Add comment"
          }
        }
      }
    },
    "/api/comments/remove": {
      "delete": {
        "tags": ["Comment"],
        "summary": "Remove comment",
        "description": "Remove comment",
        "operationId": "removeComment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "commentId": {
                    "type": "string",
                    "required": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "authCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Remove comment"
          }
        }
      }
    },
    "/api/comments/{commentId}": {
      "get": {
        "tags": ["Comment"],
        "summary": "Get comment",
        "description": "Get comment",
        "operationId": "getComment",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "Comment ID",
            "required": true,
            "example": "62d561f47233089392f88c81"
          }
        ],
        "responses": {
          "200": {
            "description": "Get comment"
          }
        }
      }
    },
...
    "/api/museums/fav-museums": {
      "get": {
        "tags": ["Museum"],
        "summary": "Get favorite museums",
        "description": "Get favorite museums",
        "operationId": "getFavMuseums",
        "security": [
          {
            "authCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of museums"
          }
        }
      }
    },
...
  },
...
}